cms-backend/requirements.txt
# FastAPI CMS Backend Requirements
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
jinja2==3.1.3
aiofiles==23.2.1

# GitHub/Gitea API
PyGithub==2.1.1
httpx==0.26.0

# Image processing
Pillow==10.2.0

# Database (optional - for sessions/users)
sqlalchemy==2.0.25
aiosqlite==0.19.0
